+++ code/core/src/node-logger/index.ts
@@ -2,7 +2,7 @@
 import npmLog from 'npmlog';
 import prettyTime from 'pretty-hrtime';
 
-export * from './prompts';
+export { prompt } from './prompts';
 
 // The default is stderr, which can cause some tools (like rush.js) to think
 // there are issues with the build: https://github.com/storybookjs/storybook/issues/14621
